requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.2
aiosignal==1.3.1
alembic==1.10.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
fastapi==0.115.4
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==0.16.3
httptools==0.6.4
httpx==0.23.1
idna==3.10
iniconfig==2.0.0

Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
propcache==0.2.0
pycparser==2.22
pydantic==2.4.0
pydantic_core==2.10.0
PyJWT==2.9.0
Pyment==0.3.3
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
python-dotenv==0.21.1
python-multipart==0.0.17
PyYAML==6.0.2
rfc3986==1.5.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.23.0
watchfiles==0.24.0
websockets==13.1
yarl==1.17.1
